+++ packages/shared/src/index.ts
@@ -85,3 +85,4 @@ export * from "./erp.js";
 export * from "./websiteBuilder.js";
 export * from "./socialPlatform.js";
 export * from "./helpdesk.js";
+export * from "./appBuilder.js";